Revert "Move the Config struct to data.go, since it can be used on Macs"
This reverts commit 28b2986d826907e72d27ed4f44712a4433a1259d.
